requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
openai>=1.30.1
python-dotenv>=1.0.1
pydantic>=2.7.1
httpx>=0.27.0
pytest>=8.1.0
pytest-asyncio>=0.23.6
